FSI: shift single field solvers to separate files
