+++ include/exadg/fluid_structure_interaction/driver.cpp
@@ -22,9 +22,7 @@
 // ExaDG
 #include <exadg/fluid_structure_interaction/acceleration_schemes/linear_algebra.h>
 #include <exadg/fluid_structure_interaction/driver.h>
-#include <exadg/grid/get_dynamic_mapping.h>
-#include <exadg/incompressible_navier_stokes/spatial_discretization/create_operator.h>
-#include <exadg/incompressible_navier_stokes/time_integration/create_time_integrator.h>
+#include <exadg/utilities/print_general_infos.h>
 #include <exadg/utilities/print_solver_results.h>
 
 namespace ExaDG
@@ -92,193 +90,6 @@ Driver<dim, Number>::setup()
   setup_interface_coupling();
 
   timer_tree.insert({"FSI", "Setup"}, timer.wall_time());
-}
-
-template<int dim, typename Number>
-void
-WrapperStructure<dim, Number>::setup(
-  std::shared_ptr<StructureFSI::ApplicationBase<dim, Number>> application,
-  MPI_Comm const                                              mpi_comm,
-  bool const                                                  is_test)
-{
-  // setup spatial operator
-  pde_operator =
-    std::make_shared<Structure::Operator<dim, Number>>(application->get_grid(),
-                                                       application->get_boundary_descriptor(),
-                                                       application->get_field_functions(),
-                                                       application->get_material_descriptor(),
-                                                       application->get_parameters(),
-                                                       "elasticity",
-                                                       mpi_comm);
-
-  // initialize matrix_free
-  matrix_free_data = std::make_shared<MatrixFreeData<dim, Number>>();
-  matrix_free_data->append(pde_operator);
-
-  matrix_free = std::make_shared<dealii::MatrixFree<dim, Number>>();
-  matrix_free->reinit(*application->get_grid()->mapping,
-                      matrix_free_data->get_dof_handler_vector(),
-                      matrix_free_data->get_constraint_vector(),
-                      matrix_free_data->get_quadrature_vector(),
-                      matrix_free_data->data);
-
-  pde_operator->setup(matrix_free, matrix_free_data);
-
-  // initialize postprocessor
-  postprocessor = application->create_postprocessor();
-  postprocessor->setup(pde_operator->get_dof_handler(), *application->get_grid()->mapping);
-
-  // initialize time integrator
-  time_integrator = std::make_shared<Structure::TimeIntGenAlpha<dim, Number>>(
-    pde_operator, postprocessor, application->get_parameters(), mpi_comm, is_test);
-
-  time_integrator->setup(application->get_parameters().restarted_simulation);
-
-  pde_operator->setup_solver();
-}
-
-template<int dim, typename Number>
-void
-WrapperFluid<dim, Number>::setup(
-  std::shared_ptr<FluidFSI::ApplicationBase<dim, Number>> application,
-  MPI_Comm const                                          mpi_comm,
-  bool const                                              is_test)
-{
-  // ALE: initialize PDE operator
-  if(application->get_parameters().mesh_movement_type == IncNS::MeshMovementType::Poisson)
-  {
-    ale_poisson_operator = std::make_shared<Poisson::Operator<dim, Number, dim>>(
-      application->get_grid(),
-      application->get_boundary_descriptor_ale_poisson(),
-      application->get_field_functions_ale_poisson(),
-      application->get_parameters_ale_poisson(),
-      "Poisson",
-      mpi_comm);
-  }
-  else if(application->get_parameters().mesh_movement_type == IncNS::MeshMovementType::Elasticity)
-  {
-    ale_elasticity_operator = std::make_shared<Structure::Operator<dim, Number>>(
-      application->get_grid(),
-      application->get_boundary_descriptor_ale_elasticity(),
-      application->get_field_functions_ale_elasticity(),
-      application->get_material_descriptor_ale_elasticity(),
-      application->get_parameters_ale_elasticity(),
-      "ale_elasticity",
-      mpi_comm);
-  }
-  else
-  {
-    AssertThrow(false, dealii::ExcMessage("not implemented."));
-  }
-
-  // ALE: initialize matrix_free_data
-  ale_matrix_free_data = std::make_shared<MatrixFreeData<dim, Number>>();
-
-  if(application->get_parameters().mesh_movement_type == IncNS::MeshMovementType::Poisson)
-  {
-    if(application->get_parameters_ale_poisson().enable_cell_based_face_loops)
-      Categorization::do_cell_based_loops(*application->get_grid()->triangulation,
-                                          ale_matrix_free_data->data);
-
-    ale_matrix_free_data->append(ale_poisson_operator);
-  }
-  else if(application->get_parameters().mesh_movement_type == IncNS::MeshMovementType::Elasticity)
-  {
-    ale_matrix_free_data->append(ale_elasticity_operator);
-  }
-  else
-  {
-    AssertThrow(false, dealii::ExcMessage("not implemented."));
-  }
-
-  // ALE: initialize matrix_free
-  ale_matrix_free = std::make_shared<dealii::MatrixFree<dim, Number>>();
-  ale_matrix_free->reinit(*application->get_grid()->mapping,
-                          ale_matrix_free_data->get_dof_handler_vector(),
-                          ale_matrix_free_data->get_constraint_vector(),
-                          ale_matrix_free_data->get_quadrature_vector(),
-                          ale_matrix_free_data->data);
-
-  // ALE: setup PDE operator and solver
-  if(application->get_parameters().mesh_movement_type == IncNS::MeshMovementType::Poisson)
-  {
-    ale_poisson_operator->setup(ale_matrix_free, ale_matrix_free_data);
-    ale_poisson_operator->setup_solver();
-  }
-  else if(application->get_parameters().mesh_movement_type == IncNS::MeshMovementType::Elasticity)
-  {
-    ale_elasticity_operator->setup(ale_matrix_free, ale_matrix_free_data);
-    ale_elasticity_operator->setup_solver();
-  }
-  else
-  {
-    AssertThrow(false, dealii::ExcMessage("not implemented."));
-  }
-
-  // ALE: create grid motion object
-  if(application->get_parameters().mesh_movement_type == IncNS::MeshMovementType::Poisson)
-  {
-    ale_grid_motion =
-      std::make_shared<GridMotionPoisson<dim, Number>>(application->get_grid()->mapping,
-                                                       ale_poisson_operator);
-  }
-  else if(application->get_parameters().mesh_movement_type == IncNS::MeshMovementType::Elasticity)
-  {
-    ale_grid_motion = std::make_shared<GridMotionElasticity<dim, Number>>(
-      application->get_grid()->mapping,
-      ale_elasticity_operator,
-      application->get_parameters_ale_elasticity());
-  }
-  else
-  {
-    AssertThrow(false, dealii::ExcMessage("not implemented."));
-  }
-
-  // initialize pde_operator
-  pde_operator = IncNS::create_operator<dim, Number>(application->get_grid(),
-                                                     ale_grid_motion,
-                                                     application->get_boundary_descriptor(),
-                                                     application->get_field_functions(),
-                                                     application->get_parameters(),
-                                                     "fluid",
-                                                     mpi_comm);
-
-  // initialize matrix_free
-  matrix_free_data = std::make_shared<MatrixFreeData<dim, Number>>();
-  matrix_free_data->append(pde_operator);
-
-  matrix_free = std::make_shared<dealii::MatrixFree<dim, Number>>();
-  if(application->get_parameters().use_cell_based_face_loops)
-    Categorization::do_cell_based_loops(*application->get_grid()->triangulation,
-                                        matrix_free_data->data);
-  std::shared_ptr<dealii::Mapping<dim> const> mapping =
-    get_dynamic_mapping<dim, Number>(application->get_grid(), ale_grid_motion);
-  matrix_free->reinit(*mapping,
-                      matrix_free_data->get_dof_handler_vector(),
-                      matrix_free_data->get_constraint_vector(),
-                      matrix_free_data->get_quadrature_vector(),
-                      matrix_free_data->data);
-
-  // setup Navier-Stokes operator
-  pde_operator->setup(matrix_free, matrix_free_data);
-
-  // setup postprocessor
-  postprocessor = application->create_postprocessor();
-  postprocessor->setup(*pde_operator);
-
-  // setup time integrator before calling setup_solvers (this is necessary since the setup
-  // of the solvers depends on quantities such as the time_step_size or gamma0!!!)
-  AssertThrow(application->get_parameters().solver_type == IncNS::SolverType::Unsteady,
-              dealii::ExcMessage("Invalid parameter in context of fluid-structure interaction."));
-
-  // initialize time_integrator
-  time_integrator = IncNS::create_time_integrator<dim, Number>(
-    pde_operator, application->get_parameters(), mpi_comm, is_test, postprocessor);
-
-  time_integrator->setup(application->get_parameters().restarted_simulation);
-
-  pde_operator->setup_solvers(time_integrator->get_scaling_factor_time_derivative_term(),
-                              time_integrator->get_velocity());
 }
 
 template<int dim, typename Number>
@@ -379,39 +190,6 @@ Driver<dim, Number>::synchronize_time_step_size() const
   // The fluid domain is the master that dictates the time step size
   structure->time_integrator->set_current_time_step_size(
     fluid->time_integrator->get_time_step_size());
-}
-
-template<int dim, typename Number>
-void
-WrapperFluid<dim, Number>::solve_ale(
-  std::shared_ptr<FluidFSI::ApplicationBase<dim, Number>> application,
-  bool const                                              is_test) const
-{
-  dealii::Timer timer;
-  timer.restart();
-
-  dealii::Timer sub_timer;
-
-  sub_timer.restart();
-  bool const print_solver_info = time_integrator->print_solver_info();
-  ale_grid_motion->update(time_integrator->get_next_time(), print_solver_info and not(is_test));
-  timer_tree->insert({"ALE", "Solve and reinit mapping"}, sub_timer.wall_time());
-
-  sub_timer.restart();
-  std::shared_ptr<dealii::Mapping<dim> const> mapping =
-    get_dynamic_mapping<dim, Number>(application->get_grid(), ale_grid_motion);
-  matrix_free->update_mapping(*mapping);
-  timer_tree->insert({"ALE", "Update matrix-free"}, sub_timer.wall_time());
-
-  sub_timer.restart();
-  pde_operator->update_after_grid_motion();
-  timer_tree->insert({"ALE", "Update operator"}, sub_timer.wall_time());
-
-  sub_timer.restart();
-  time_integrator->ale_update();
-  timer_tree->insert({"ALE", "Update time integrator"}, sub_timer.wall_time());
-
-  timer_tree->insert({"ALE"}, timer.wall_time());
 }
 
 template<int dim, typename Number>
